[global] 이슈 #72-#80 버그 수정 및 코드 개선

+++ src/main/kotlin/team/nongchun/hororog/domain/budget/dto/BudgetMonthlySummaryResponse.kt
@@ -2,8 +2,8 @@ package team.nongchun.hororog.domain.budget.dto
 
 data class BudgetMonthlySummaryResponse(
     val month: String,
-    val totalBudget: Int,
-    val usedBudget: Int,
-    val remaining: Int,
+    val totalBudget: Long,
+    val usedBudget: Long,
+    val remaining: Long,
     val currency: String = "KRW",
 )
